requirements.txt
paho-mqtt==1.5.1
Adafruit-DHT==1.4.0
getmac==0.8.2
influxdb==5.3.1
python-dotenv==0.19.2